Makefile clean fixed. Added \o and \a commands in main.tex

+++ Makefile
@@ -10,4 +10,4 @@ view: build
 	zathura bin/main.pdf &
 
 clean:
-	rm -r bin/
+	if [ -d bin/ ]; then rm -r bin/; fi
